Add GPRGD support for binary variables
The dictionary returned by DataUtil.dummy_encoder_helper
now has an additional field for binary variables,
which are noncontinuous variables that only take values 0 or 1.
In other words, boolean variables or enums which take 2 values.

While these could be dummy-encoded as well, it's wasteful and unnecessary.
It is necessary for GPRGD's constraint helper to adjust binary variables
back to 0/1 values after each step of gradient descent.

Resolves #137